==> code/1-original-data-prep.ipynb <==
import pandas as pd
import numpy as np
import os



with open("../data/original/data-def.txt", "w", encoding="utf-8") as a:
    for path, subdirs, files in os.walk(r'../data/original/'):
       for filename in files:
        f = os.path.join(path, filename)
        a.write(filename + os.linesep)
        df = pd.read_table(f, sep=';')
        for listitem in df.columns:
            a.write('%s\n' % listitem)
        a.write('\n')



comments = pd.DataFrame(columns=['name', 'comment', 'column_nam'])
# comments

### Leveranser av motorbensin, dieselbränsle, etanol och eldningsolja till slutgiltiga förbrukare, 1000 m3 efter region, varuslag och är

df1 = pd.read_table('../data/original/000000B6_20221201-162102.csv', sep=';')
df2 = pd.read_table('../data/original/000000B6_20221201-162137.csv', sep=';')
df3 = pd.read_table('../data/original/000000B6_20221201-162146.csv', sep=';')

df1 = df1.append(df2).append(df3)
# df1

print(len(df1), len(df1.drop_duplicates()))
filename = 'leveranser-till-slutlsiga-forbrukare-per-ar.csv'
df1.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Leveranser av motorbensin, dieselbränsle, etanol och eldningsolja till slutgiltiga förbrukare, 1000 m3 efter region, varuslag och är'

comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df1.columns.to_list()}, ignore_index=True)


### Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är
df2 = pd.read_table('../data/original/000000BQ_20221201-141152.csv', sep=';')
# df2
print(len(df2), len(df2.drop_duplicates()))

filename = 'elproduktion-per-ar.csv'
df2.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df2.columns.to_list()}, ignore_index=True)


### Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är
df3 = pd.read_table('../data/original/000000BR_20221201-145212.csv', sep=';')
# df3
print(len(df3), len(df3.drop_duplicates()))

filename = 'fjarvarmproduktion-per-ar.csv'
df3.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df3.columns.to_list()}, ignore_index=True)


# Slutanvändning (MWh) efter region, förbrukarkategori, bränsletyp och är
df4 = pd.read_table('../data/original/000000BS_20221201-145418.csv', sep=';')
# df4
print(len(df4), len(df4.drop_duplicates()))

filename = 'slutanvanding-per-ar.csv'
df4.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Slutanvändning (MWh) efter region, förbrukarkategori, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df4.columns.to_list()}, ignore_index=True)


# Elanvändning i Sverige, GWh efter användningsområde, elomröde och är
df5 = pd.read_table('../data/original/000000W8_20221201-162326.csv', sep=';')
# df5
print(len(df5), len(df5.drop_duplicates()))

filename = 'elvanding-efer-anvandingsomrade-per-ar.csv'
df5.to_csv('../data/grouped-data/'+filename)

comment = 'Elanvändning i Sverige, GWh efter användningsområde, elomröde och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df5.columns.to_list()}, ignore_index=True)


# Utleveranser av brånslen efter brånsleslag, tabellinnehåll och månad
df6 = pd.read_table('../data/original/000001D4_20221201-160300.csv', sep=';')
# df6
print(len(df6), len(df6.drop_duplicates()))

filename = 'utleveranser-per-ar.csv'
df6.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Utleveranser av brånslen efter brånsleslag, tabellinnehåll och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df6.columns.to_list()}, ignore_index=True)


# Utleveranser av biodrivmedel. Volym i m3 (om inte annat anges) efter varuslag och månad
df7 = pd.read_table('../data/original/000001H5_20221201-160349.csv', sep=';')
# df7
print(len(df7), len(df7.drop_duplicates()))

filename = 'utleveranser-av-biodrivmedel-per-ar.csv'
df7.to_csv('../data/grouped-data/' + filename)

comment = 'Utleveranser av biodrivmedel. Volym i m3 (om inte annat anges) efter varuslag och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df7.columns.to_list()}, ignore_index=True)


# Leveranser av bränsle, kubikmeter efter region, bränsletyp, färbrukarkategori och är
df8 = pd.read_table('../data/original/000001SF_20221201-162027.csv', sep=';')
df9 = pd.read_table('../data/original/000003TP_20221201-161952.csv', sep=';')
df8 = df8.append(df9)
# df8
print(len(df8), len(df8.drop_duplicates()))

filename = 'leveranser-av-bransle-per-ar.csv'
df8.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Leveranser av bränsle, kubikmeter efter region, bränsletyp, färbrukarkategori och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df8.columns.to_list()}, ignore_index=True)


# bränsleanvändning efter bränsletyp, näringsgren SNI 2007 och kvartal
df9 = pd.read_table('../data/original/000002UO_20221201-154955.csv', sep=';')

# df9
print(len(df9), len(df9.drop_duplicates()))

filename = 'bransleanvandning-after-bransletyp-per-kvartal.csv'
df9.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränsleanvändning efter bränsletyp, näringsgren SNI 2007 och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df9.columns.to_list()}, ignore_index=True)


# Elproduktion efter produktionsslag och månad
df10 = pd.read_table('../data/original/000004DF_20221201-161918.csv', sep=';')
# df10
print(len(df10), len(df10.drop_duplicates()))

filename = 'elproduktion-per-manad.csv'
df10.to_csv('../data/grouped-data/' + filename)

comment = 'Elproduktion efter produktionsslag och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df10.columns.to_list()}, ignore_index=True)


# bränslefärbrukning fär elproduktion, TJ efter produktionsslag, bränsletyp och är
df11 = pd.read_table('../data/original/000005GI_20221201-162309.csv', sep=';')
# df11
print(len(df11), len(df11.drop_duplicates()))

filename = 'branslefarbrukning-per-ar.csv'
df11.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränslefärbrukning fär elproduktion, TJ efter produktionsslag, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df11.columns.to_list()}, ignore_index=True)


# Bruttoäverfäring av el till och från andra lånder, GWh efter land och månad
df12 = pd.read_table('../data/original/000005J8_20221201-161937.csv', sep=';')
# df12
print(len(df12), len(df12.drop_duplicates()))

filename = 'bruttoaverfaring-per-manad.csv'
df12.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Bruttoäverfäring av el till och från andra lånder, GWh efter land och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df12.columns.to_list()}, ignore_index=True)


# bränsleanvändning efter bränsletyp, produktion, anlåggning och kvartal
df13 = pd.read_table('../data/original/000005JP_20221201-155035.csv', sep=';')

# df13
print(len(df13), len(df13.drop_duplicates()))

filename = 'bransleanvandning-per-kvartal.csv'
df13.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränsleanvändning efter bränsletyp, produktion, anlåggning och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df13.columns.to_list()}, ignore_index=True)


# Uttagspunkter efter användningsområde, elomröde och är
df14 = pd.read_table('../data/original/000006HL_20221201-162233.csv', sep=';')
# df14
print(len(df14), len(df14.drop_duplicates()))

filename = 'uttagspunkter-per-ar.csv'
df14.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Uttagspunkter efter användningsområde, elomröde och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df14.columns.to_list()}, ignore_index=True)


# Fjärrkyla efter tabellinnehåll och är
df15 = pd.read_table('../data/original/000006HN_20221201-162532.csv', sep=';')
# df15
print(len(df15), len(df15.drop_duplicates()))

filename = 'fjarrkyla-per-ar.csv'
df15.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Fjärrkyla efter tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df15.columns.to_list()}, ignore_index=True)


# anlåggningar, installerad effekt samt bruttoproduktion efter elomröde, kraftslag, tabellinnehåll och är
df16 = pd.read_table('../data/original/000006HQ_20221201-162204.csv', sep=';')
# df16
print(len(df16), len(df16.drop_duplicates()))

filename = 'anlaggningar-per-ar.csv'
df16.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'anlåggningar, installerad effekt samt bruttoproduktion efter elomröde, kraftslag, tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df16.columns.to_list()}, ignore_index=True)


# Omsatt värme efter kategori, tabellinnehåll och är
df17 = pd.read_table('../data/original/000006HU_20221201-162816.csv', sep=';')
# df17
print(len(df17), len(df17.drop_duplicates()))

filename = 'omsatt-per-ar.csv'
df17.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Omsatt värme efter kategori, tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df17.columns.to_list()}, ignore_index=True)


# Nettoproduktion av el, bränsleanvändning samt verkningsgrad efter kraftslag, tabellinnehåll och är
df18 = pd.read_table('../data/original/000006I6_20221201-162555.csv', sep=';')
# df18
print(len(df18), len(df18.drop_duplicates()))

filename = 'nettoproduktion-per-ar.csv'
df18.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Nettoproduktion av el, bränsleanvändning samt verkningsgrad efter kraftslag, tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df18.columns.to_list()}, ignore_index=True)


# Värmeleveranser till slutliga färbrukare efter färbrukarkategori, tabellinnehåll och är
df19 = pd.read_table('../data/original/000006I7_20221201-162846.csv', sep=';')
# df19
print(len(df19), len(df19.drop_duplicates()))

filename = 'varmeleveranser-per-ar.csv'
df19.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Värmeleveranser till slutliga färbrukare efter färbrukarkategori, tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df19.columns.to_list()}, ignore_index=True)


# Tillfärsel av el-energi, GWh efter produktionsslag, tabellinnehåll och är
df21 = pd.read_table('../data/original/EN0105A2_20221201-162902.csv', sep=';')
df22 = pd.read_table('../data/original/EN0105A3_20221201-162915.csv', sep=';')
# df21 = df21.append(df22); print("1", len(df21), len(df21.drop_duplicates()))
df23 = pd.read_table('../data/original/EN0105B0_20221201-162929.csv', sep=';')
# df21 = df21.append(df23); print("2", len(df21), len(df21.drop_duplicates()))
df24 = pd.read_table('../data/original/EN0105B1_20221201-163134.csv', sep=';')
# df21 = df21.append(df24); print("3", len(df21), len(df21.drop_duplicates()))
df25 = pd.read_table('../data/original/EN0105A1_20221201-162428.csv', sep=';')
# df21 = df21.append(df25); print("4", len(df21), len(df21.drop_duplicates()))
df29 = pd.read_table('../data/original/EN0105B1_20221201-163134.csv', sep=';')
df21 = df21.append(df22).append(df23).append(df24).append(df25).append(df29)

# df21
print(len(df21), len(df21.drop_duplicates()))

filename = 'varmeleveranser-per-ar.csv'
df21.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Tillfärsel av el-energi, GWh efter produktionsslag, tabellinnehåll och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df21.columns.to_list()}, ignore_index=True)


# bränslefärbrukning fär produktion av ånga och hetvatten, TJ efter produktionsslag, bränsletyp och är
df22 = pd.read_table('../data/original/EN0105A7_20221201-162251.csv', sep=';')
# df22
print(len(df22), len(df22.drop_duplicates()))

filename = 'branslefarbrukning-far-produktion-per-ar.csv'
df22.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränslefärbrukning fär produktion av ånga och hetvatten, TJ efter produktionsslag, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df22.columns.to_list()}, ignore_index=True)


# bränsleanvändning, 1000 toe efter näringsgren SNI92, bränsletyp och kvartal
df23 = pd.read_table('../data/original/EN0106A1_20221201-155056.csv', sep=';')

# df24
print(len(df23), len(df23.drop_duplicates()))

filename = 'branslefarbrukning-after-naringsgren-per-kvartal.csv'
df23.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränsleanvändning, 1000 toe efter näringsgren SNI92, bränsletyp och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df23.columns.to_list()}, ignore_index=True)


# bränsleanvändning, terajoule (TJ) efter näringsgren SNI 2007, bränsletyp och kvartal
df24 = pd.read_table('../data/original/EN0106B3_20221201-154909.csv', sep=';')

# df24
print(len(df24), len(df24.drop_duplicates()))

filename = 'branslefarbrukning-SNI2007-per-kvartal.csv'
df24.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränsleanvändning, terajoule (TJ) efter näringsgren SNI 2007, bränsletyp och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df24.columns.to_list()}, ignore_index=True)


# bränsleanvändning, terajoule (TJ) efter näringsgren SNI92, bränsletyp och kvartal
df25 = pd.read_table('../data/original/EN0106C0_20221201-154936.csv', sep=';')

# df25
print(len(df25), len(df25.drop_duplicates()))

filename = 'branslefarbrukning-SNI92-per-kvartal.csv'
df25.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'bränsleanvändning, terajoule (TJ) efter näringsgren SNI92, bränsletyp och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df25.columns.to_list()}, ignore_index=True)


# Leveranser av flytande bränslen samt gasol, m3 och ton efter varuslag, färbrukarkategori och månad
df26 = pd.read_table('../data/original/EN0107A2_20221201-160423.csv', sep=';')
df27 = pd.read_table('../data/original/EN0107A3_20221201-160406.csv', sep=';')
df26 = df26.append(df27)

# df26
print(len(df26), len(df26.drop_duplicates()))

filename = 'leveranser-per-manad.csv'
df26.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Leveranser av flytande bränslen samt gasol, m3 och ton efter varuslag, färbrukarkategori och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df26.columns.to_list()}, ignore_index=True)


# Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI92 och månad
df27 = pd.read_table('../data/original/EN0108A2_20221201-160711.csv', sep=';')

# df27
print(len(df27), len(df27.drop_duplicates()))

filename = 'elfarbrukning-per-manad.csv'
df27.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI92 och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df27.columns.to_list()}, ignore_index=True)


# Eltillfärsel netto, GWh efter produktionsslag och månad
df28 = pd.read_table('../data/original/EN0108A3_20221201-160443.csv', sep=';')

# df28
print(len(df28), len(df28.drop_duplicates()))

filename = 'eltillfarsel-per-manad.csv'
df28.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Eltillfärsel netto, GWh efter produktionsslag och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df28.columns.to_list()}, ignore_index=True)


# Elanvändning, GWh efter användningsområde och månad
df29 = pd.read_table('../data/original/EN0108A4_20221201-160536.csv', sep=';')

# df29
print(len(df29), len(df29.drop_duplicates()))

filename = 'eltillfarsel-per-manad.csv'
df29.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elanvändning, GWh efter användningsområde och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df29.columns.to_list()}, ignore_index=True)


# Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI69 och månad
df30 = pd.read_table('../data/original/EN0108A5_20221201-161826.csv', sep=';')

# df30
print(len(df30), len(df30.drop_duplicates()))

filename = 'elfarbrukning-mineral-per-manad.csv'
df30.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI69 och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df30.columns.to_list()}, ignore_index=True)


# Byten av elleverantär samt beräknad bytesvolym efter kundkategori, tabellinnehåll och månad
df31 = pd.read_table('../data/original/EN0108A6_20221201-161859.csv', sep=';')

# df31
print(len(df31), len(df31.drop_duplicates()))

filename = 'byten-elleverantar-per-manad.csv'
df31.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Byten av elleverantär samt beräknad bytesvolym efter kundkategori, tabellinnehåll och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df31.columns.to_list()}, ignore_index=True)


### Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI 2007 och månad
df32 = pd.read_table('../data/original/EN0108B1_20221201-160553.csv', sep=';')

# df32
print(len(df32), len(df32.drop_duplicates()))

filename = 'elfarbrukning-SNI2007-per-manad.csv'
df32.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elfärbrukning inom utvinning av mineral, tillverkningsindustri, GWh efter näringsgren SNI 2007 och månad'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df32.columns.to_list()}, ignore_index=True)


### Kvartalsvisa energibalanser, TJ efter energibärare, energiflöden och kvartal
df33 = pd.read_table('../data/original/EN0201A3_20221201-160229.csv', sep=';')
df34 = pd.read_table('../data/original/EN0201A5_20221201-160034.csv', sep=';')
# df33 = df33.append(df34); print("1", len(df33), len(df33.drop_duplicates()))
df35 = pd.read_table('../data/original/EN0201A5_20221201-160053.csv', sep=';')
# df33 = df33.append(df35); print("2", len(df33), len(df33.drop_duplicates()))
df37 = pd.read_table('../data/original/EN0201A8_20221201-155420.csv', sep=';')
# df33 = df33.append(df37); print("4", len(df33), len(df33.drop_duplicates()))

df33 = df33.append(df34).append(df35).append(df37)

# df33
print(len(df33), len(df33.drop_duplicates()))

filename = 'energibalanser-per-kvartal.csv'
df33.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Kvartalsvisa energibalanser, TJ efter energibärare, energiflöden och kvartal'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df33.columns.to_list()}, ignore_index=True)


### Energidata (MWh) efter region, kategori, energityp och år
df34 = pd.read_table('../data/original/EN0203AB_20221201-145612.csv', sep=';')
df35 = pd.read_table('../data/original/EN0203AB_20221201-145644.csv', sep=';')
# df34 = df34.append(df35); print("1", len(df34), len(df34.drop_duplicates()))
df37 = pd.read_table('../data/original/EN0203AB_20221201-152954.csv', sep=';')
# df34 = df34.append(df37); print("3", len(df34), len(df34.drop_duplicates()))
df38 = pd.read_table('../data/original/EN0203AB_20221201-153033.csv', sep=';')
# df34 = df34.append(df38); print("4", len(df34), len(df34.drop_duplicates()))
df39 = pd.read_table('../data/original/EN0203AB_20221201-153355.csv', sep=';')
# df34 = df34.append(df39); print("5", len(df34), len(df34.drop_duplicates()))
df40 = pd.read_table('../data/original/EN0203AB_20221201-153431.csv', sep=';')
# df34 = df34.append(df40); print("6", len(df34), len(df34.drop_duplicates()))
df41 = pd.read_table('../data/original/EN0203AB_20221201-153513.csv', sep=';')
# df34 = df34.append(df41); print("7", len(df34), len(df34.drop_duplicates()))
df42 = pd.read_table('../data/original/EN0203AB_20221201-153616.csv', sep=';')
# df34 = df34.append(df42); print("8", len(df34), len(df34.drop_duplicates()))
df43 = pd.read_table('../data/original/EN0203AB_20221201-153651.csv', sep=';')
# df34 = df34.append(df43); print("9", len(df34), len(df34.drop_duplicates()))
df45 = pd.read_table('../data/original/EN0203AB_20221201-154043.csv', sep=';')
# df34 = df34.append(df45); print("11", len(df34), len(df34.drop_duplicates()))
df46 = pd.read_table('../data/original/EN0203AB_20221201-154126.csv', sep=';')
# df34 = df34.append(df46); print("12", len(df34), len(df34.drop_duplicates()))
df47 = pd.read_table('../data/original/EN0203AB_20221201-154140.csv', sep=';')
# df34 = df34.append(df47); print("13", len(df34), len(df34.drop_duplicates()))
df50 = pd.read_table('../data/original/EN0203AB_20221201-154207.csv', sep=';')
# df34 = df34.append(df50); print("16", len(df34), len(df34.drop_duplicates()))
df51 = pd.read_table('../data/original/EN0203AB_20221201-154216.csv', sep=';')
# df34 = df34.append(df51); print("17", len(df34), len(df34.drop_duplicates()))
df53 = pd.read_table('../data/original/EN0203AB_20221201-154441.csv', sep=';')
# df34 = df34.append(df53); print("19", len(df34), len(df34.drop_duplicates()))
df54 = pd.read_table('../data/original/EN0203AB_20221201-154452.csv', sep=';')
# df34 = df34.append(df54); print("20", len(df34), len(df34.drop_duplicates()))
df55 = pd.read_table('../data/original/EN0203AB_20221201-154500.csv', sep=';')
# df34 = df34.append(df55); print("21", len(df34), len(df34.drop_duplicates()))
df56 = pd.read_table('../data/original/EN0203AB_20221201-154508.csv', sep=';')
# df34 = df34.append(df56); print("22", len(df34), len(df34.drop_duplicates()))
df57 = pd.read_table('../data/original/EN0203AB_20221201-154520.csv', sep=';')
# df34 = df34.append(df57); print("23", len(df34), len(df34.drop_duplicates()))
df59 = pd.read_table('../data/original/EN0203AB_20221201-154539.csv', sep=';')
# df34 = df34.append(df59); print("25", len(df34), len(df34.drop_duplicates()))
df60 = pd.read_table('../data/original/EN0203AB_20221201-154550.csv', sep=';')
# df34 = df34.append(df60); print("26", len(df34), len(df34.drop_duplicates()))
df61 = pd.read_table('../data/original/EN0203AB_20221201-154559.csv', sep=';')
# df34 = df34.append(df61); print("27", len(df34), len(df34.drop_duplicates()))
df62 = pd.read_table('../data/original/EN0203AB_20221201-154609.csv', sep=';')
# df34 = df34.append(df62); print("28", len(df34), len(df34.drop_duplicates()))
df63 = pd.read_table('../data/original/EN0203AB_20221201-154620.csv', sep=';')
# df34 = df34.append(df63); print("29", len(df34), len(df34.drop_duplicates()))
df64 = pd.read_table('../data/original/EN0203AB_20221201-154632.csv', sep=';')
# df34 = df34.append(df64); print("30", len(df34), len(df34.drop_duplicates()))
df65 = pd.read_table('../data/original/EN0203AB_20221201-154644.csv', sep=';')
# df34 = df34.append(df65); print("31", len(df34), len(df34.drop_duplicates()))
df66 = pd.read_table('../data/original/EN0203AB_20221201-154653.csv', sep=';')
# df34 = df34.append(df66); print("32", len(df34), len(df34.drop_duplicates()))
df68 = pd.read_table('../data/original/EN0203AB_20221201-154745.csv', sep=';')
# df34 = df34.append(df68); print("34", len(df34), len(df34.drop_duplicates()))
df69 = pd.read_table('../data/original/EN0203AB_20221201-154756.csv', sep=';')
# df34 = df34.append(df69); print("35", len(df34), len(df34.drop_duplicates()))
df70 = pd.read_table('../data/original/EN0203AB_20221201-154804.csv', sep=';')
# df34 = df34.append(df70); print("36", len(df34), len(df34.drop_duplicates()))
df74 = pd.read_table('../data/original/EN0203AB_20221201-154835.csv', sep=';')
# df34 = df34.append(df74); print("40", len(df34), len(df34.drop_duplicates()))


df34 = df34.append(df35).append(df37).append(df38).append(df39
        ).append(df40).append(df41).append(df42).append(df43).append(df45
        ).append(df46).append(df47).append(df50).append(df51).append(df53
        ).append(df54).append(df55).append(df56).append(df57).append(df59
        ).append(df60).append(df61).append(df62).append(df63).append(df64
        ).append(df65).append(df66).append(df68).append(df69).append(df70).append(df74)

# df34
print(len(df34), len(df34.drop_duplicates()))

filename = 'energidata-per-ar.csv'
df34.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Energidata (MWh) efter region, kategori, energityp och år'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df34.columns.to_list()}, ignore_index=True)


### Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är
df35 = pd.read_table('../data/original/EN0203AC_20221201-141231.csv', sep=';')
df36 = pd.read_table('../data/original/EN0203AC_20221201-141255.csv', sep=';')
df37 = pd.read_table('../data/original/EN0203AC_20221201-145005.csv', sep=';')
df38 = pd.read_table('../data/original/EN0203AC_20221201-145050.csv', sep=';')

df35 = df35.append(df36).append(df37).append(df38)


# df35
print(len(df35), len(df35.drop_duplicates()))

filename = 'fjarrvarmeproduktion-per-ar.csv'
df35.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df35.columns.to_list()}, ignore_index=True)


### Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är
df36 = pd.read_table('../data/original/EN0203AD_20221201-141125.csv', sep=';')
df36
print(len(df36), len(df36.drop_duplicates()))

filename = 'elproduktion-&-bransleanvandning-per-ar.csv'
df36.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df36.columns.to_list()}, ignore_index=True)


### Slutanvändning (MWh) efter region, färbrukarkategori, bränsletyp och är
df37 = pd.read_table('../data/original/EN0203AE_20221201-145253.csv', sep=';')
df38 = pd.read_table('../data/original/EN0203AE_20221201-145332.csv', sep=';')
df39 = pd.read_table('../data/original/EN0203AB_20221201-154043.csv', sep=';')

df37 = df37.append(df38).append(df39)

# df37
print(len(df37), len(df37.drop_duplicates()))

filename = 'slutanvandning-per-ar.csv'
df37.to_csv('../data/grouped-data/' + filename, index=False)

comment = 'Slutanvändning (MWh) efter region, färbrukarkategori, bränsletyp och är'
comments = comments.append({'name': filename, 'comment': comment, 'column_nam': df37.columns.to_list()}, ignore_index=True)


# Write document with filenames desc and list of columns

with open("../data/grouped-data/clean-data-explanations.txt", "w", encoding="utf-8") as a:
    for i in comments.index:
        # print(comments.comment[i])
        a.write(comments.name[i] + os.linesep)
        a.write(comments.comment[i] + os.linesep)
        for listitem in comments.column_nam[i]:
            # print(listitem)
            a.write('%s\n' % listitem)
        a.write(os.linesep)


# # Write document with filenames desc and list of columns

# with open("clean-data-explanations2.txt", "w", encoding="utf-8") as a:
#     for path, subdirs, files in os.walk(r'../data/', ):
#         for filename, i in zip(files, comments.index):
#             f = os.path.join(path, filename)
#             a.write(filename + ': ')
#             # print(comments.comment[i])
#             a.write(comments.comment[i] + os.linesep)
#             df = pd.read_table(f, sep=',', low_memory=False,
#                                on_bad_lines='skip')
#             # print(f)
#             for listitem in df.columns:
#                 # print(listitem)
#                 a.write('%s\n' % listitem)
#             # a.write(os.linesep)
#             a.write(os.linesep)
