==> code/3-data-file-prep.ipynb <==
# Förnybar/icke förnybar gruppering
## Bränsletyp

|typ|icke-förnybar/förnybar|bransletyp namn|
|---|----------------------|---------------|
|flytande bransletyp|icke-förnybar|motorbensin|
|flytande bransletyp|icke-förnybar|bensin|
|flytande bransletyp|icke-förnybar|diesel|
|flytande bransletyp|icke-förnybar|diesel/EO1|
|flytande bransletyp|icke-förnybar|stenkol|
|flytande bransletyp|icke-förnybar|fotogen|
|flytande bransletyp|icke-förnybar|naturgas|
|flytande bransletyp|icke-förnybar|varav  diesel|
|flytande bransletyp|icke-förnybar|dieselbr�nsle|
|flytande bransletyp|icke-förnybar|E85|
|flytande bransletyp|icke-förnybar|avfall|
|flytande bransletyp|icke-förnybar|propan och butan|
|flytande bransletyp|icke-förnybar|torv och torvbriketter| <!--peat and peat briquettes-->
|flytande bransletyp|icke-förnybar|stadsgas och koksugnsgas|
|flytande bransletyp|icke-förnybar|masugnsgas inkl. LD-gas|
|flytande bransletyp|icke-förnybar|eldningsolja nr 1|
|flytande bransletyp|icke-förnybar|eldningsolja 1|
|flytande bransletyp|icke-förnybar|varav eldningsolja 1|
|flytande bransletyp|icke-förnybar|eldningsolja 2|
|flytande bransletyp|icke-förnybar|eldningsolja nr 2|
|flytande bransletyp|icke-förnybar|eldningsolja nr 2 och 3|
|flytande bransletyp|icke-förnybar|eldningsolja 2-6|
|flytande bransletyp|icke-förnybar|eldningsolja 3-6|
|flytande bransletyp|icke-förnybar|eldningsolja nr 3-5|
|flytande bransletyp|icke-förnybar|eldningsolja nr 1 d�rav ej SNI 05-33, 35|
|flytande bransletyp|icke-förnybar|eldningsolja nr 2 inkl. WRD|
|flytande bransletyp|icke-förnybar|eldningsolja nr 2 inkl. WRD d�rav ej SNI 05-33|
|flytande bransletyp|icke-förnybar|Eldningsolja nr 2-5|
|flytande bransletyp|icke-förnybar|Eldningsolja nr 2-5 d�rav ej SNI 05-33, 35|
|flytande bransletyp|icke-förnybar|eldningsolja nr 3-6|
|flytande bransletyp|icke-förnybar|eldningsolja nr 3-6 d�rav ej SNI 05-33, 35|
|flytande bransletyp|icke-förnybar|eldningsolja nr 4|
|flytande bransletyp|icke-förnybar|eldningsolja nr 5 och d�r�ver|
|flytande bransletyp|icke-förnybar|flytande (icke f�rnybara)|
|fast bransletyp|icke-förnybar|fast (icke f�rnybara)|
|gas bransletyp|icke-förnybar|gas (icke f�rnybara)|
|--------------|-------------|---------------------|
|flytande bransletyp|förnybar|FAME|
|flytande bransletyp|förnybar|f�r�dlade tr�dbr�nslen|
|flytande bransletyp|förnybar|tr�dbr�nslen exklusive f�r�dlade tr�dbr�nslen|
|flytande bransletyp|förnybar|HVO|
|flytande bransletyp|förnybar|etanol|
|flytande bransletyp|förnybar|biogas|
|flytande bransletyp|förnybar|svartlutar, tall- och beckolja|
|flytande bransletyp|förnybar|flytande (f�rnybara)|
|fast bransletyp|förnybar|fast (f�rnybara)|
|gas bransletyp|förnybar|gas (f�rnybara)|
|--------------|-------------|---------------------|
|okand bransletyp|ökand|NAN|
|okand bransletyp|ökand|elproduktion|
|okand bransletyp|ökand|el|
|okand bransletyp|ökand|fj�rrv�rme|
|okand bransletyp|ökand|annat br�nsle|
|--------------|-------------|---------------------|
|okand bransletyp|summa|summa br�nsle och drivmedel|
|okand bransletyp|summa|summa produktionss�tt|
|okand bransletyp|summa|summa f�rbrukarkategori|


## Produktionssatt
|typ|icke-förnybar/förnybar|bransletyp namn|
|---|----------------------|---------------|
||icke-förnybar|kraftv�rmeverk + industriellt mottryck|
||icke-förnybar|�vrig v�rmekraft (k�rnkraft, kondenskraft o.dyl.)|
||icke-förnybar|kraftv�rmeverk|
||icke-förnybar|frist�ende v�rmeverk|
||icke-förnybar|spillv�rme|
||icke-förnybar|r�kgaskondens|
||icke-förnybar|kraftv�rme-v�rmeverk, kraftv�rmedrift|
||icke-förnybar|kraftv�rme-v�rmeverk, enbart v�rmeproduktion|
||icke-förnybar|frist�ende v�rmeverk|
|---|----------------------|---------------|
||förnybar|vattenkraft|
||förnybar|vindkraft|
|---|----------------------|---------------|
||kärnkraft|�vrig v�rmekraft (k�rnkraft, kondenskraft o.dyl.)|
|---|----------------------|---------------|
|okand bransletyp|okand|elpannor (1)|
|okand bransletyp|okand|v�rmepumpar (2)|
|okand bransletyp|summa|summa|
|okand bransletyp|summa|summa br�nsletyp|



## Kategori
|typ|icke-förnybar/förnybar|bransletyp namn|
|---|----------------------|---------------|
||icke-förnybar|3.4 insats V�rmekraftverk (ej k�rn-)|
||icke-förnybar|3.5 insats Indust.mottrycksanl.|
||icke-förnybar|3.6.1 insats Kraftv�rmeverk, fj�rrv�rmeprod.|
||icke-förnybar|3.6.2 insats Kraftv�rmeverk, elprod.|
||icke-förnybar|3.7 insats Frist�ende v�rmeverk|
||icke-förnybar|4.4 omvandlat V�rmekraftverk (ej k�rn-)|
||icke-förnybar|4.5 omvandlat Industr.mottrycksanl.|
||icke-förnybar|4.6 omvandlat Kraftv�rmeverk|
||icke-förnybar|4.6.2 omvandlat Spec. Kraftv.spillv�rme|
||icke-förnybar|4.7 omvandlat Frist�ende v�rmeverk|
||icke-förnybar|4.7.2 omvandlat Spec.Frist.v�rmev. spillv�rme|
||icke-förnybar|5.4 AnvEnSektor V�rmekraftverk (ej k�rn-)|
||icke-förnybar|5.5 AnvEnSektor Industr.mottrycksanl�ggning|
||icke-förnybar|5.6 AnvEnSektor Kraftv�rmeverk|
||icke-förnybar|5.7 AnvEnSektor Frist�ende v�rmeverk|
||icke-förnybar|7 �verf�ringsf�rluster|
||icke-förnybar|9.2 slutanv. Industri, byggverks.|
|---|----------------------|---------------|
||kärnkraft|5.3 AnvEnSektor K�rnkraftverk|
||kärnkraft|3.3 insats K�rnkraftverk|
||kärnkraft|4.3 omvandlat K�rnkraftverk|
|---|----------------------|---------------|
||förnybar|9.1 slutanv. Jordbruk,skogsbruk,fiske|
||förnybar|3.1 insats Vattenkraftstationer|
||förnybar|3.2 insats Vindkraftverk|
||förnybar|4.1 omvandlat Vattenkraftstationer|
||förnybar|4.2 omvandlat Vindkraftverk|
||förnybar|5.1 AnvEnSektor Vattenkraftstationer|
||förnybar|5.2 AnvEnSektor Vindkraftverk|
|---|----------------------|---------------|
|okand bransletyp|okand|5 anv i energisektor totalt|
|okand bransletyp|okand|3.6.1.1 insats Spec. Fj�rrv�rme m. v�rmepumpar|
|okand bransletyp|okand|3.7.1 insats Spec.Frist�ende v�rmev. v�rmepumpar|
|okand bransletyp|okand|4 omvandlat totalt|
|okand bransletyp|okand|4.6.1omvandlat Spec. Kraftv.v�rmepumpar|
|okand bransletyp|okand|4.7.1 omvandlat Spec.Frist.v�rmev. v�rmepumpar|
|okand bransletyp|okand|5.6.1 AnvEnSektor Spec.Kraftv.v�rmepumpar|
|okand bransletyp|okand|5.7.1 AnvEnSektor Frist.v�rmev.v�rmepumpar|
|okand bransletyp|okand|9 slutlig anv�ndning totalt|
|okand bransletyp|okand|2 bruttotillf�rsel|
|okand bransletyp|okand|3 insatt f�r omvandling totalt|
|okand bransletyp|okand|9.3 slutanv. Offentlig verksamhet|
|okand bransletyp|okand|9.4 slutanv. Transporter|
|okand bransletyp|okand|9.5 slutanv. �vriga tj�nster|
|okand bransletyp|okand|9.6 slutanv. Hush�ll|
|okand bransletyp|okand|9.6.1 slutanv. Spec Hush�ll sm�hus|
|okand bransletyp|okand|9.6.1.1 slutanv. Spec Hush�ll eluppv.sm�hus|
|okand bransletyp|okand|9.6.2 slutanv. Spec Hush�ll flerbostadshus        |
|okand bransletyp|okand|9.6.3 slutanv. Spec Hush�ll fritidshus|

## Energityp
|typ|icke-förnybar/förnybar|bransletyp namn|
|---|----------------------|---------------|
||icke-förnybar|stenkol|
||icke-förnybar|koks|
||icke-förnybar|bensin|
||icke-förnybar|diesel|
||icke-förnybar|eldningsolja 1|
||icke-förnybar|eldningsolja>1|
||icke-förnybar|gasol|
||icke-förnybar|naturgas|
||icke-förnybar|torv|
||icke-förnybar|avlutar|
||icke-förnybar|avfall|
|---|----------------------|---------------|
||förnybar|tr�br�nsle|
|---|----------------------|---------------|
|okand bransletyp|okand|�vrigt|
|okand bransletyp|okand|fj�rrv�rme|
|okand bransletyp|okand|el|
|okand bransletyp|okand|total energi|


<!-- ## Förbrukarkategori

jordbruk, skogsbruk, fiske
industri
el- och v�rmeverk
offentlig f�rvaltning
enbostadshus
flerbostadshus
�vriga fastigheter
v�gtransporter
bygg- och transporttj�nster, luftfart, j�rnv�g...
totalt
bygg, luftfart, j�rnv�g, bunkring
slutanv. jordbruk,skogsbruk,fiske
slutanv. industri, byggverks.
slutanv. offentlig verksamhet
slutanv. transporter
slutanv. �vriga tj�nster
slutanv. sm�hus
slutanv. flerbostadshus
slutanv. fritidshus
summa br�nsletyp
NaN
slutanv. jordbruk,skogsbruk,fiske
slutanv. industri, byggverks.
slutanv. offentlig verksamhet
slutanv. transporter
slutanv. �vriga tj�nster
slutanv. sm�hus
slutanv. flerbostadshus
slutanv. fritidshus
summa br�nsletyp -->

import pandas as pd
import numpy as np
import os

## Dataframe carying info about the files for creating documentation
datasets = pd.DataFrame(columns=['orig1', 'comment1', 'orig2', 'comment2','filename', 'column_nam'])


## Groupings of reneable/non-renewable sources
ickeFornybar_bransletyp = pd.Series(["motorbensin", "bensin", "diesel", "diesel/EO1", "stenkol", "fotogen", "naturgas", 
                                    "varav  diesel", "dieselbr�nsle", "E85", "avfall", "propan och butan", 
                                    "torv och torvbriketter ", "stadsgas och koksugnsgas", "masugnsgas inkl. LD-gas", 
                                    "eldningsolja nr 1", "eldningsolja 1", "varav eldningsolja 1", "eldningsolja 2", 
                                    "eldningsolja nr 2", "eldningsolja nr 2 och 3", "eldningsolja 2-6", "eldningsolja 3-6",
                                    "eldningsolja nr 3-5", "eldningsolja nr 1 d�rav ej SNI 05-33, 35", 
                                    "eldningsolja nr 2 inkl. WRD", "eldningsolja nr 2 inkl. WRD d�rav ej SNI 05-33", 
                                    "Eldningsolja nr 2-5", "Eldningsolja nr 2-5 d�rav ej SNI 05-33, 35", 
                                    "eldningsolja nr 3-6", "eldningsolja nr 3-6 d�rav ej SNI 05-33, 35", 
                                    "eldningsolja nr 4", "eldningsolja nr 5 och d�r�ver", "flytande(icke f�rnybara)", 
                                    "fast(icke f�rnybara)", "gas(icke f�rnybara)"])
fornybar_bransletyp = pd.Series(["FAME", "f�r�dlade tr�dbr�nslen", "tr�dbr�nslen exklusive f�r�dlade tr�dbr�nslen", "HVO",
                                "etanol", "biogas", "svartlutar, tall - och beckolja", "flytande(f�rnybara)", "fast(f�rnybara)", "gas(f�rnybara)"])
okand_bransletyp = pd.Series(["NAN", "summa produktionss�tt", "summa f�rbrukarkategori",
                             "elproduktion", "el", "fj�rrv�rme", "annat br�nsle", "summa br�nsle och drivmedel"])


## Produktionssatt

ickeFornybar_produktionssatt = pd.Series(["kraftv�rmeverk + industriellt mottryck", "�vrig v�rmekraft (k�rnkraft, kondenskraft o.dyl.)", 
                                        "kraftv�rmeverk", "frist�ende v�rmeverk", "spillv�rme", "r�kgaskondens", "kraftv�rme-v�rmeverk, kraftv�rmedrift", 
                                        "kraftv�rme-v�rmeverk, enbart v�rmeproduktion", "frist�ende v�rmeverk"])
fornybar_productionssatt = pd.Series(["vattenkraft", "vindkraft"])
karnkraft_produktionssatt = pd.Series(["�vrig v�rmekraft(k�rnkraft, kondenskraft o.dyl.)"])
okand_produktionssatt = pd.Series(["elpannor (1)", "v�rmepumpar (2)"])

summa_produktionssatt = pd.Series(["summa", "summa br�nsletyp"])


## Kategori
ickFornybar_kategori = pd.Series(["3.4 insats V�rmekraftverk (ej k�rn-)", "3.5 insats Indust.mottrycksanl.", 
                                "3.6.1 insats Kraftv�rmeverk, fj�rrv�rmeprod.", "3.6.2 insats Kraftv�rmeverk, elprod.", 
                                "3.7 insats Frist�ende v�rmeverk", "4.4 omvandlat V�rmekraftverk (ej k�rn-)", 
                                "4.5 omvandlat Industr.mottrycksanl.", "4.6 omvandlat Kraftv�rmeverk", 
                                "4.6.2 omvandlat Spec. Kraftv.spillv�rme", "4.7 omvandlat Frist�ende v�rmeverk", 
                                "4.7.2 omvandlat Spec.Frist.v�rmev. spillv�rme", "5.4 AnvEnSektor V�rmekraftverk (ej k�rn-)", 
                                "5.5 AnvEnSektor Industr.mottrycksanl�ggning", "5.6 AnvEnSektor Kraftv�rmeverk", 
                                "5.7 AnvEnSektor Frist�ende v�rmeverk", "7 �verf�ringsf�rluster", 
                                "9.2 slutanv. Industri, byggverks."])
karnkraft_kategori = pd.Series(["5.3 AnvEnSektor K�rnkraftverk", "3.3 insats K�rnkraftverk", 
                                "4.3 omvandlat K�rnkraftverk"])
fornybar_kategori = pd.Series(["9.1 slutanv. Jordbruk,skogsbruk,fiske", "3.1 insats Vattenkraftstationer", 
                            "3.2 insats Vindkraftverk", "4.1 omvandlat Vattenkraftstationer", 
                            "4.2 omvandlat Vindkraftverk", "5.1 AnvEnSektor Vattenkraftstationer", "5.2 AnvEnSektor Vindkraftverk"])
okand_kategori = pd.Series(["5 anv i energisektor totalt", "3.6.1.1 insats Spec. Fj�rrv�rme m. v�rmepumpar", 
                            "3.7.1 insats Spec.Frist�ende v�rmev. v�rmepumpar", "4 omvandlat totalt", "4.6.1omvandlat Spec. Kraftv.v�rmepumpar", 
                            "4.7.1 omvandlat Spec.Frist.v�rmev. v�rmepumpar", "5.6.1 AnvEnSektor Spec.Kraftv.v�rmepumpar", 
                            "5.7.1 AnvEnSektor Frist.v�rmev.v�rmepumpar", "9 slutlig anv�ndning totalt", "2 bruttotillf�rsel", 
                            "3 insatt f�r omvandling totalt", "9.3 slutanv. Offentlig verksamhet", "9.4 slutanv. Transporter", 
                            "9.5 slutanv. �vriga tj�nster", "9.6 slutanv. Hush�ll", "9.6.1 slutanv. Spec Hush�ll sm�hus", 
                            "9.6.1.1 slutanv. Spec Hush�ll eluppv.sm�hus", "9.6.2 slutanv. Spec Hush�ll flerbostadshus", 
                            "9.6.3 slutanv. Spec Hush�ll fritidshus"])

## Energityp
ickeFornybar_energityp = pd.Series(["stenkol", "koks", "bensin", "diesel", "eldningsolja 1", "eldningsolja>1", "gasol", 
                                    "naturgas", "torv", "avlutar", "avfall"])
fornybar_energityp = pd.Series(["tr�br�nsle"])
okand_energityp = pd.Series(["�vrigt", "fj�rrv�rme", "el", "total energi"])

# Leveranser av motorbensin, dieselbränsle, etanol och eldningsolja till slutgiltiga förbrukare, 1000 m3 efter region, varuslag och är

df1 = pd.read_table('../data/grouped-data/leveranser-till-slutlsiga-forbrukare-per-ar.csv', sep=',')

df1 = df1.melt(id_vars=['region', 'varuslag'], var_name='ar', value_name='m3')

df1 = df1.rename(columns={'varuslag': 'bransletyp'})

print(df1.columns)
# print('region\n', df1.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df1.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df1.ar.drop_duplicates(keep='first').reset_index(drop=True))


# print(df1.describe)

# print(df1.dtypes)

orig1 = 'leveranser-till-slutlsiga-forbrukare-per-ar.csv'
comment1 = 'Leveranser av motorbensin, dieselbränsle, etanol och eldningsolja till slutgiltiga förbrukare, 1000 m3 efter region, varuslag och är'


# Leveranser av bränsle, kubikmeter efter region, bränsletyp, färbrukarkategori och är

df8 = pd.read_table('../data/grouped-data/leveranser-av-bransle-per-ar.csv', sep=',')

# print(df8.columns)
df8.rename(columns={'br�nsletyp': 'bransletyp', 'f�rbrukarkategori': 'forbrukarkategori'}, inplace=True)

df8 = df8.melt(id_vars=['region', 'forbrukarkategori', 'bransletyp'], var_name='ar', value_name='m3')

# print('region\n', df8.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('forbrukarkategori\n', df8.forbrukarkategori.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df8.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df8.ar.drop_duplicates(keep='first').reset_index(drop=True))
# print(df8.describe)

# print(df8.dtypes)

orig2 = 'leveranser-av-bransle-per-ar.csv'
comment2 = 'Leveranser av bränsle, kubikmeter efter region, bränsletyp, färbrukarkategori och är'


# merge 1 & 8 on region, bransletyp, ar

df1 = pd.concat([df1, df8])

df1 = df1.replace('..', 0)

df1['fornybar_bransletyp'] = np.where(df1.bransletyp.isin(fornybar_bransletyp), "fornybar", np.where(df1.bransletyp.isin(ickeFornybar_bransletyp), "icke-fornybar", "okand"))


filename = 'leveranser-flytande-bransle-region-ar.csv'
df1.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': orig2, 'comment2': comment2, 'filename':filename, 'column_nam': df1.columns.to_list()}, ignore_index=True)


# Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är

df2 = pd.read_table('../data/grouped-data/elproduktion-per-ar.csv', sep=',')

# print(df3.columns)
df2.rename(columns = {'produktionss�tt':'produktionssatt', 'br�nsletyp':'bransletyp'}, inplace = True)

df2 = df2.melt(id_vars=['region', 'produktionssatt', 'bransletyp'], var_name='ar', value_name='MWh')

# print('region\n', df2.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('produktionssatt\n', df2.produktionssatt.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df2.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df2.ar.drop_duplicates(keep='first').reset_index(drop=True))


# print(df2.describe)

# print(df2.dtypes)

orig1 = 'elproduktion-per-ar.csv'
comment1 = 'Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'


# Elproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är

df35 = pd.read_table('../data/grouped-data/elproduktion-&-bransleanvandning-per-ar.csv', sep=',')

# print(df35.columns)

df35.rename(columns={'produktionss�tt': 'produktionssatt', 'br�nsletyp': 'bransletyp'}, inplace=True)

df35 = df35.melt(id_vars=['region', 'produktionssatt', 'bransletyp'], var_name='ar', value_name='MWh')

# print('region\n', df35.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('produktionssatt\n', df35.produktionssatt.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df35.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df35.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df35.describe)

# print(df35.dtypes)

orig2 = 'elproduktion-&-bransleanvandning-per-ar.csv'
comment2 = 'Elproduktion och bränsleanvändning(MWh) efter region, produktionssätt, bränsletyp och är'


# merge 2 & 35 on 'region', 'produktionssatt', 'bransletyp', ar
df2 = pd.concat([df2, df35])

df2 = df2.replace('..', 0)

df2['fornybar_bransletyp'] = np.where(df2.bransletyp.isin(fornybar_bransletyp), "fornybar", np.where(df2.bransletyp.isin(ickeFornybar_bransletyp), "icke-fornybar", "okand"))
df2['fornybar_produktionssatt'] = np.where(df2.produktionssatt.isin(ickeFornybar_produktionssatt), "fornybar",
                                    np.where(df2.produktionssatt.isin(fornybar_productionssatt), "icke-fornybar",
                                        np.where(df2.produktionssatt.isin(karnkraft_produktionssatt), "karnkraft", 
                                            np.where(df2.produktionssatt.isin(summa_produktionssatt), "summa", "okand"))))


filename = 'elproduction-bransleanvandning-region-produktionssatt-bransletyp-ar.csv'
df2.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': orig2, 'comment2': comment2, 'filename': filename, 'column_nam': df2.columns.to_list()}, ignore_index=True)


# Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är

df3 = pd.read_table('../data/grouped-data/fjarvarmproduktion-per-ar.csv', sep=',')

# print(df3.columns)
df3.rename(columns={'produktionss�tt': 'produktionssatt', 'br�nsletyp': 'bransletyp'}, inplace=True)

df3 = df3.melt(id_vars=['region', 'produktionssatt', 'bransletyp'], var_name='ar', value_name='MWh')

# print('region\n', df3.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('produktionssatt\n', df3.produktionssatt.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df3.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df3.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df3.describe)

# print(df3.dtypes)

orig1 = 'fjarvarmproduktion-per-ar.csv'
comment1 = 'Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'


# Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är

df34 = pd.read_table('../data/grouped-data/fjarrvarmeproduktion-per-ar.csv', sep=',', low_memory=False)

# print(df34.columns)

df34.rename(columns={'produktionss�tt': 'produktionssatt', 'br�nsletyp': 'bransletyp'}, inplace=True)

df34 = df34.melt(id_vars=['region', 'produktionssatt', 'bransletyp'], var_name='ar', value_name='MWh')


# print('region\n', df34.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('produktionssatt\n', df34.produktionssatt.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df34.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df34.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df34.describe)

# print(df34.dtypes)

orig2 = 'fjarrvarmeproduktion-per-ar.csv'
comment2 = 'Fjärrvärmeproduktion och bränsleanvändning (MWh) efter region, produktionssätt, bränsletyp och är'


# merge 3 & 34 on 'region', 'produktionssatt', 'bransletyp', 'ar'
df3 = pd.concat([df3, df34])

df3 = df3.replace('..', 0)

df3['fornybar_bransletyp'] = np.where(df3.bransletyp.isin(fornybar_bransletyp), "fornybar", np.where(
    df3.bransletyp.isin(ickeFornybar_bransletyp), "icke-fornybar", "okand"))
df3['fornybar_produktionssatt'] = np.where(df3.produktionssatt.isin(ickeFornybar_produktionssatt), "fornybar",
                                           np.where(df3.produktionssatt.isin(fornybar_productionssatt), "icke-fornybar",
                                                    np.where(df3.produktionssatt.isin(karnkraft_produktionssatt), "karnkraft",
                                                             np.where(df3.produktionssatt.isin(summa_produktionssatt), "summa", "okand"))))


filename = 'fjarrvarmeproduktion-region-produktionssatt-bransletyp-ar.csv'
df3.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': orig2, 'comment2': comment2, 'filename': filename, 'column_nam': df3.columns.to_list()}, ignore_index=True)


# Slutanvändning (MWh) efter region, förbrukarkategori, bränsletyp och är

df4 = pd.read_table('../data/grouped-data/slutanvanding-per-ar.csv', sep=',')

# print(df4.columns)

df4.rename(columns={'f�rbrukarkategori': 'forbrukarkategori', 'br�nsletyp': 'bransletyp'}, inplace=True)


df4 = df4.melt(id_vars=['region', 'forbrukarkategori', 'bransletyp'], var_name='ar', value_name='MWh')

# print('region\n', df4.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('forbrukarkategori\n', df4.forbrukarkategori.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df4.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df4.ar.drop_duplicates(keep='first').reset_index(drop=True))


# print(df4.describe)

# print(df4.dtypes)

orig1 = 'slutanvanding-per-ar.csv'
comment1 = 'Slutanvändning (MWh) efter region, förbrukarkategori, bränsletyp och är'


# Slutanvändning (MWh) efter region, färbrukarkategori, bränsletyp och är
# fuel type & region & year
df36 = pd.read_table('../data/grouped-data/slutanvandning-per-ar.csv', sep=',')

# print(df36.columns)

df36.rename(columns={'f�rbrukarkategori': 'forbrukarkategori', 'br�nsletyp': 'bransletyp'}, inplace=True)
df36 = df36.melt(id_vars=['region', 'forbrukarkategori', 'bransletyp', 'kategori', 'energityp'], var_name='ar', value_name='MWh')

# print('region\n', df36.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('forbrukarkategori\n', df36.forbrukarkategori.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df36.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df36.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df36.describe)

# print(df36.dtypes)

orig2 = 'slutanvandning-per-ar.csv'
comment2 = 'Slutanvändning (MWh) efter region, färbrukarkategori, bränsletyp och är'


# merge 4 & 36 on 'region', 'forbrukarkategori', 'bransletyp', 'ar'
df4 = pd.concat([df4, df36])

df4 = df4.replace('..', 0)

df4['fornybar_bransletyp'] = np.where(df4.bransletyp.isin(fornybar_bransletyp), "fornybar", np.where(
    df4.bransletyp.isin(ickeFornybar_bransletyp), "icke-fornybar", "okand"))
    


filename = 'slutanvandning-region-forbrukarkategori-bransletyp-kategori-energityp-ar.csv'
df4.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': orig2, 'comment2': comment2, 'filename': filename, 'column_nam': df4.columns.to_list()}, ignore_index=True)


# Bränslefärbrukning fär produktion av ånga och hetvatten, TJ efter produktionsslag, bränsletyp och är
# fuel type & year
df21 = pd.read_table('../data/grouped-data/branslefarbrukning-far-produktion-per-ar.csv', sep=',')

# print(df21.columns)

df21.rename(columns={'br�nsletyp': 'bransletyp'}, inplace=True)

df21 = df21.melt(id_vars=['produktionsslag', 'bransletyp'], var_name='ar', value_name='branslefarbrukning')

# print('produktionsslag\n', df21.produktionsslag.drop_duplicates(keep='first').reset_index(drop=True))
# print('bransletyp\n', df21.bransletyp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df21.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df21.describe)

# print(df21.dtypes)

df21['fornybar_bransletyp'] = np.where(df21.bransletyp.isin(fornybar_bransletyp), "fornybar", np.where(
    df21.bransletyp.isin(ickeFornybar_bransletyp), "icke-fornybar", "okand"))


orig1 = 'branslefarbrukning-far-produktion-per-ar'
comment1 = 'Bränslefärbrukning fär produktion av ånga och hetvatten, TJ efter produktionsslag, bränsletyp och är'

filename = 'branslefarbrukning-far-produktion-per-ar.csv'
df21.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': 'NaN', 'comment2': 'NaN', 'filename': filename, 'column_nam': df21.columns.to_list()}, ignore_index=True)


# Energidata (MWh) efter region, kategori, energityp och år

df33 = pd.read_table('../data/grouped-data/energidata-per-ar.csv', sep=',')

# print(df33.columns)

df33 = df33.melt(id_vars=['region', 'kategori', 'energityp'], var_name='ar', value_name='MWh')

# print('region\n', df33.region.drop_duplicates(keep='first').reset_index(drop=True))
# print('kategori\n', df33.kategori.drop_duplicates(keep='first').reset_index(drop=True))
# print('energityp\n', df33.energityp.drop_duplicates(keep='first').reset_index(drop=True))
# print('ar\n', df33.ar.drop_duplicates(keep='first').reset_index(drop=True))

# print(df33.describe)

# print(df33.dtypes)

df33['fornybar_kategori'] = np.where(df33.kategori.isin(ickFornybar_kategori), "fornybar",
                                     np.where(df33.kategori.isin(fornybar_kategori), "icke-fornybar",
                                                    np.where(df33.kategori.isin(karnkraft_kategori), "karnkraft", "okand")))
df33['fornybar_energityp'] = np.where(df33.energityp.isin(ickeFornybar_energityp), "fornybar",
                                      np.where(df33.energityp.isin(fornybar_energityp), "icke-fornybar", "okand"))


orig1 = 'energidata-per-ar.csv'
comment1 = 'Energidata (MWh) efter region, kategori, energityp och år'

filename = 'energidata-per-ar.csv'
df33.to_csv('../data/to-use/' + filename, index=False)
datasets = datasets.append({'orig1': orig1, 'comment1': comment1, 'orig2': 'NaN', 'comment2': 'NaN', 'filename': filename, 'column_nam': df33.columns.to_list()}, ignore_index=True)


# Write document with filenames desc and list of columns

with open("../data/to-use/to-use-data-explanations.txt", "w", encoding="utf-8") as a:
    for i in datasets.index:
        # print(comments.comment[i])
        a.write(datasets.filename[i] + ':' + os.linesep)
        a.write(datasets.orig1[i] + ' ')
        a.write('(' + datasets.comment1[i] + ')' + os.linesep)
        a.write(datasets.orig2[i] + ' ')
        a.write('(' + datasets.comment2[i] + ')' + os.linesep)
        for listitem in datasets.column_nam[i]:
            # print(listitem)
            a.write('%s\n' % listitem)
        a.write(os.linesep)
